fix(release): fix merge back PR creation
When loading existing PRs via GitHub API, the `head` branch filter needs
to be prefixed with the branch owner, e.g. `picter:release` instead of
just `release`. Otherwise it seems to be ignored, so any PRs with the
defined `base` are loaded.
This caused the release action to re-use existing PRs to master for the
merge back PR (by overwriting their title and description).